Add dialog box after sale transaction for summary and tax implication calculations

+++ init_db.py
@@ -6,8 +6,8 @@ Date TEXT NOT NULL,
 Transaction_Type TEXT NOT NULL,
 Fund_Name TEXT NOT NULL,
 Scheme_Code TEXT NOT NULL,
-Units INTEGER NOT NULL,
-Amount INTEGER NOT NULL)"""
+Units REAL NOT NULL,
+Amount REAL NOT NULL)"""
 
 with sqlite3.Connection("retirement_manager.db") as conn:
     conn.execute(mutual_fund_table)
